pyproject.toml
[project]
name = "distri-of-population"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "OrkWard", email = "orkward_o@outlook.com"},
]
dependencies = [
    "geopandas>=0.14.3",
    "ipykernel>=6.29.3",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pyright]
venvPath = "."
venv = ".venv"
